feat: refine not found handling

+++ src/controllers/banner,controller.ts
@@ -2,7 +2,7 @@ import { Response } from "express"
 import { IPaginationQuery, IReqUser } from "../utils/interfaces"
 import response from "../utils/response";
 import BannerModel, { bannerDAO, TypeBanner } from "../models/banner.model";
-import { FilterQuery } from "mongoose";
+import { FilterQuery, isValidObjectId } from "mongoose";
 
 
 export default {
@@ -54,6 +54,9 @@ export default {
     async findOne (req: IReqUser, res: Response) {
                  try {
             const { id } = req.params;
+            if (!isValidObjectId(id)) {
+        return response.notFound(res, "failed find one a ticket");
+      }
             const result = await BannerModel.findById(id);
             if (!result) {
                 return response.notFound(res, "failed to find this banner");
@@ -67,6 +70,9 @@ export default {
     async update (req: IReqUser, res: Response) {
          try {
             const { id } = req.params;
+            if (!isValidObjectId) {
+                return response.notFound(res, "failed to update this banner");
+                }
             const result = await BannerModel.findByIdAndUpdate(id, req.body, {new: true,});
             response.success(res, result, "success to update banner");
         }catch (error) {
@@ -76,6 +82,9 @@ export default {
     async remove (req: IReqUser, res: Response) {
         try {
             const { id } = req.params;
+            if (!isValidObjectId) {
+                return response.notFound(res, "failed to remove this banner");
+                }
             const result = await BannerModel.findByIdAndDelete(id, {new: true,});
             response.success(res, result, "success to remove banner");
         }catch (error) {
