+++ src/controllers/ticket.controller.ts
@@ -55,6 +55,9 @@ export default {
     async findOne(req: IReqUser, res: Response) {
         try {
             const { id } = req.params;
+            if (!isValidObjectId(id)) {
+                return response.notFound(res, "failed to find one ticket");
+            }
             const result = await TicketModel.findById(id);
             if (!result) {
                             return response.notFound(res, "failed to find one ticket");
@@ -67,6 +70,9 @@ export default {
     async update(req: IReqUser, res: Response) {
         try {
             const { id } = req.params;
+            if (!isValidObjectId(id)) {
+                return response.notFound(res, "failed to update a ticket");
+            }
             const result = await TicketModel.findByIdAndUpdate(id, req.body,{new: true,});
             response.success(res, result, "success update a ticket");
     } catch (error) {
@@ -76,6 +82,9 @@ export default {
     async remove(req: IReqUser, res: Response) {
         try {
             const { id } = req.params;
+            if (!isValidObjectId(id)) {
+                return response.error(res, null, "failed to remove ticket");
+            }
             const result = await TicketModel.findByIdAndDelete(id,{new: true,});
             response.success(res, result, "success remove a ticket");
     } catch (error) {
